Use namui only if feature enabled
